requirements.txt
httpx[http2]
dnspython
pyyaml
rich